+++ src/token.rs
@@ -2,8 +2,11 @@
 
 use crate::engine::{
     Engine, KEYWORD_DEBUG, KEYWORD_EVAL, KEYWORD_FN_PTR, KEYWORD_FN_PTR_CALL, KEYWORD_FN_PTR_CURRY,
-    KEYWORD_IS_SHARED, KEYWORD_PRINT, KEYWORD_THIS, KEYWORD_TYPE_OF,
+    KEYWORD_IS_DEF_FN, KEYWORD_IS_DEF_VAR, KEYWORD_PRINT, KEYWORD_THIS, KEYWORD_TYPE_OF,
 };
+
+#[cfg(not(feature = "no_closure"))]
+use crate::engine::KEYWORD_IS_SHARED;
 
 use crate::error::LexError;
 use crate::parser::INT;
@@ -507,9 +510,11 @@ impl Token {
             | "await" | "yield" => Reserved(syntax.into()),
 
             KEYWORD_PRINT | KEYWORD_DEBUG | KEYWORD_TYPE_OF | KEYWORD_EVAL | KEYWORD_FN_PTR
-            | KEYWORD_FN_PTR_CALL | KEYWORD_FN_PTR_CURRY | KEYWORD_IS_SHARED | KEYWORD_THIS => {
-                Reserved(syntax.into())
-            }
+            | KEYWORD_FN_PTR_CALL | KEYWORD_FN_PTR_CURRY | KEYWORD_IS_DEF_VAR
+            | KEYWORD_IS_DEF_FN | KEYWORD_THIS => Reserved(syntax.into()),
+
+            #[cfg(not(feature = "no_closure"))]
+            KEYWORD_IS_SHARED => Reserved(syntax.into()),
 
             _ => return None,
         })
@@ -1455,7 +1460,9 @@ pub fn is_keyword_function(name: &str) -> bool {
         #[cfg(not(feature = "no_closure"))]
         KEYWORD_IS_SHARED => true,
         KEYWORD_PRINT | KEYWORD_DEBUG | KEYWORD_TYPE_OF | KEYWORD_EVAL | KEYWORD_FN_PTR
-        | KEYWORD_FN_PTR_CALL | KEYWORD_FN_PTR_CURRY => true,
+        | KEYWORD_FN_PTR_CALL | KEYWORD_FN_PTR_CURRY | KEYWORD_IS_DEF_VAR | KEYWORD_IS_DEF_FN => {
+            true
+        }
         _ => false,
     }
 }
@@ -1465,7 +1472,8 @@ pub fn is_keyword_function(name: &str) -> bool {
 #[inline(always)]
 pub fn can_override_keyword(name: &str) -> bool {
     match name {
-        KEYWORD_PRINT | KEYWORD_DEBUG | KEYWORD_TYPE_OF | KEYWORD_EVAL | KEYWORD_FN_PTR => true,
+        KEYWORD_PRINT | KEYWORD_DEBUG | KEYWORD_TYPE_OF | KEYWORD_EVAL | KEYWORD_FN_PTR
+        | KEYWORD_IS_DEF_VAR | KEYWORD_IS_DEF_FN => true,
         _ => false,
     }
 }